Add ready method and autoFetch/autoFetchRelated
The ready method acts like jQuery's ready method. If it is called and
passed a function, it calls that function when the model as been
completely initialized (either now or later depending on model.isReady).

The autoFetch/autoFetchRelated attributes allow the coder to indicate
that a model should be fetched on load without having to pass options.